Restructured HTTP extractors. Removed Features.cc as it was a stupid file name. Code moved to better named files.
