+++ plugin/CMakeLists.txt
@@ -15,7 +15,6 @@ add_library(plugin SHARED
 	src/Ex_Base.cc
 	src/Ex_HTTP.cc
 	src/Modifier.cc
-	src/Feature.cc
 	src/Machinery.cc
         src/Rxp.cc
 	src/ts_util.cc
