+++ plugin/src/Context.cc
@@ -188,7 +188,7 @@ Context& Context::commit(Feature &feature) {
       memcpy(span, *fv);
       fv->_direct_p = false;
       fv->_literal_p = true;
-      *fv = span.view(); // update view to be the localized copy.
+      *fv = span.view(); // update full to be the localized copy.
     } else if (fv->data() == _arena->remnant().data()) { // it's in transient memory, finalize it.
       _arena->alloc(fv->size());
       fv->_literal_p = true;
